requirements.txt
streamlit==1.35.0
google-generativeai>=0.5.0
PyMuPDF